requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.12.1
twilio==8.10.0
websockets==12.0
deepgram-sdk==3.2.7
google-generativeai==0.3.2
elevenlabs==0.2.26
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
